requirements.txt
langchain==0.2.5
langchain-core==0.2.9
langchain-aws==0.1.7
langgraph==0.0.69
playwright==1.44.0
python-dotenv==1.0.1
ipython==8.26.0
ipywidgets==8.1.3
nest_asyncio==1.6.0
boto3>=1.34.0
